+++ shared/infra/http/middleware/recover_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/royhq/go-play-app/infra/http/middleware"
+	"github.com/royhq/go-play-app/shared/infra/http/middleware"
 )
 
 func TestWithRecover(t *testing.T) {
