+++ use_cases/ping/endpoint.go
