+++ use_cases/users/create/endpoint_errors_test.go
@@ -15,7 +15,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/royhq/go-play-app/features/users/create"
+	"github.com/royhq/go-play-app/use_cases/users/create"
 )
 
 func TestEndpointErrorHandler_HandleError(t *testing.T) {
