+++ use_cases/users/create/endpoint_test.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/royhq/go-play-app/features/users/create"
+	"github.com/royhq/go-play-app/use_cases/users/create"
 )
 
 func TestEndpointHandler_ServeHTTP(t *testing.T) {
