+++ use_cases/users/get/domain.go
@@ -1,0 +1,4 @@
+package get
+
+type User struct {
+}
