+++ use_cases/users/get/endpoint.go
@@ -1,0 +1,1 @@
+package get
