+++ use_cases/users/get/query.go
@@ -1,0 +1,4 @@
+package get
+
+type Query struct {
+}
