+++ use_cases/users/get/storage.go
@@ -1,0 +1,1 @@
+package get
